dns_sd(macOS): move to its own file, tighten unsafe, address review
- Move the shared-connection backend out of dns.rs into
  dns_jsc/dns_sd.rs.
- SharedConnection methods now take &mut self; unsafe blocks wrap only
  the libsystem_dnssd calls and the individual raw-pointer accesses that
  need them, instead of whole method bodies.
- Initialize the internal request's query state before starting the
  subordinate, and only log the dns_sd cache miss when dns_sd actually
  took the lookup.
- Trim paragraph-length comments.
- Tests: cover IP-literal / scoped-IPv6 / hints routing on macOS, and
  compare the unref-retry child's output as a sorted array so duplicate
  lines are not masked.
