+++ src/js/node/net.ts
@@ -4089,8 +4089,7 @@ function initSocketHandle(self) {
   const handle = self._handle;
   if (handle) {
     handle[owner_symbol] = self;
-    // A fresh handle (e.g. the autoSelectFamily retry's) must inherit an
-    // unref() the user already applied.
+    // A fresh handle (e.g. an autoSelectFamily retry) inherits a prior unref().
     if (self[kUserUnrefed]) handle.unref?.();
   }
 }
