+++ src/runtime/dns_jsc/dns.rs
@@ -119,777 +119,13 @@ bun_output::declare_scope!(DNSResolver, visible);
 const IANA_DNS_PORT: i32 = 53;
 
 // ──────────────────────────────────────────────────────────────────────────
-// dns_sd (macOS DNSServiceGetAddrInfo via mDNSResponder)
-//
-// libinfo's `getaddrinfo_async_start` is a thin wrapper that dispatches a
-// blocking `si_addrinfo` onto the libdispatch overcommit root queue: one
-// parked worker thread per in-flight lookup (Libinfo si_module.c
-// `si_async_call`). Under the hood that blocking call is itself a kqueue-
-// poll of `DNSServiceQueryRecord` against mDNSResponder. Talking to the
-// daemon directly via `DNSServiceGetAddrInfo` over one shared connection
-// removes the thread wrapper: one Unix-domain socket, one kqueue knote,
-// zero extra threads regardless of in-flight count.
+// dns_sd (macOS): DNSServiceGetAddrInfo over one shared mDNSResponder
+// connection — no per-lookup threads (unlike libinfo's async wrapper).
 // ──────────────────────────────────────────────────────────────────────────
 
 #[cfg(target_os = "macos")]
-pub(crate) mod dns_sd {
-    use super::*;
-
-    pub(crate) type DNSServiceRef = *mut c_void;
-    type DNSServiceFlags = u32;
-    type DNSServiceErrorType = i32;
-    pub(crate) type DNSServiceProtocol = u32;
-
-    pub(crate) const FLAGS_MORE_COMING: DNSServiceFlags = 0x1;
-    pub(crate) const FLAGS_ADD: DNSServiceFlags = 0x2;
-    const FLAGS_RETURN_INTERMEDIATES: DNSServiceFlags = 0x1000;
-    const FLAGS_SHARE_CONNECTION: DNSServiceFlags = 0x4000;
-    const FLAGS_SUPPRESS_UNUSABLE: DNSServiceFlags = 0x8000;
-    const FLAGS_TIMEOUT: DNSServiceFlags = 0x10000;
-
-    pub(crate) const PROTOCOL_IPV4: DNSServiceProtocol = 0x01;
-    pub(crate) const PROTOCOL_IPV6: DNSServiceProtocol = 0x02;
-
-    pub(crate) const ERR_NO_ERROR: DNSServiceErrorType = 0;
-    const ERR_NO_SUCH_NAME: DNSServiceErrorType = -65538;
-    const ERR_NO_MEMORY: DNSServiceErrorType = -65539;
-    const ERR_REFUSED: DNSServiceErrorType = -65553;
-    pub(crate) const ERR_NO_SUCH_RECORD: DNSServiceErrorType = -65554;
-    const ERR_SERVICE_NOT_RUNNING: DNSServiceErrorType = -65563;
-    const ERR_NO_ROUTER: DNSServiceErrorType = -65566;
-    pub(crate) const ERR_TIMEOUT: DNSServiceErrorType = -65568;
-    const ERR_DEFUNCT_CONNECTION: DNSServiceErrorType = -65569;
-    const ERR_POLICY_DENIED: DNSServiceErrorType = -65570;
-    const ERR_NOT_PERMITTED: DNSServiceErrorType = -65571;
-
-    type GetAddrInfoReply = unsafe extern "C" fn(
-        sd_ref: DNSServiceRef,
-        flags: DNSServiceFlags,
-        interface_index: u32,
-        error_code: DNSServiceErrorType,
-        hostname: *const c_char,
-        address: *const Sockaddr,
-        ttl: u32,
-        context: *mut c_void,
-    );
-
-    // libsystem_dnssd.dylib is part of the libSystem umbrella and always linked.
-    unsafe extern "C" {
-        fn DNSServiceCreateConnection(sd_ref: *mut DNSServiceRef) -> DNSServiceErrorType;
-        fn DNSServiceRefSockFD(sd_ref: DNSServiceRef) -> c_int;
-        fn DNSServiceProcessResult(sd_ref: DNSServiceRef) -> DNSServiceErrorType;
-        fn DNSServiceRefDeallocate(sd_ref: DNSServiceRef);
-        fn DNSServiceGetAddrInfo(
-            sd_ref: *mut DNSServiceRef,
-            flags: DNSServiceFlags,
-            interface_index: u32,
-            protocol: DNSServiceProtocol,
-            hostname: *const c_char,
-            callback: GetAddrInfoReply,
-            context: *mut c_void,
-        ) -> DNSServiceErrorType;
-    }
-
-    /// Map a DNSServiceErrorType to an EAI_* code so existing error paths
-    /// (`c_ares::Error::init_eai`, `after_result`) see what they expect.
-    pub(crate) fn to_eai(err: DNSServiceErrorType) -> c_int {
-        match err {
-            ERR_NO_ERROR => 0,
-            ERR_NO_SUCH_NAME | ERR_NO_SUCH_RECORD => libc::EAI_NONAME,
-            ERR_TIMEOUT | ERR_NO_ROUTER | ERR_DEFUNCT_CONNECTION | ERR_SERVICE_NOT_RUNNING => {
-                libc::EAI_AGAIN
-            }
-            ERR_NO_MEMORY => libc::EAI_MEMORY,
-            ERR_POLICY_DENIED | ERR_NOT_PERMITTED | ERR_REFUSED => libc::EAI_FAIL,
-            _ => libc::EAI_FAIL,
-        }
-    }
-
-    pub(crate) fn protocol_for_family(family: bun_dns::Family) -> DNSServiceProtocol {
-        match family {
-            bun_dns::Family::Inet => PROTOCOL_IPV4,
-            bun_dns::Family::Inet6 => PROTOCOL_IPV6,
-            // Both explicitly, not 0: per-family NoSuchRecord/Timeout replies
-            // from `ReturnIntermediates` are the only per-query completion
-            // signal on a shared connection (MoreComing is collective), so the
-            // requested-family set must be known up front.
-            _ => PROTOCOL_IPV4 | PROTOCOL_IPV6,
-        }
-    }
-
-    pub(crate) fn protocol_for_hints(hints: &AddrInfo) -> DNSServiceProtocol {
-        match hints.ai_family {
-            f if f == netc::AF_INET => PROTOCOL_IPV4,
-            f if f == netc::AF_INET6 => PROTOCOL_IPV6,
-            _ => PROTOCOL_IPV4 | PROTOCOL_IPV6,
-        }
-    }
-
-    /// Names getaddrinfo treats as literals rather than DNS queries: IP
-    /// addresses (including `inet_aton` numbers-and-dots shorthand such as
-    /// `127.1`) and scoped IPv6 (`fe80::1%en0`). The daemon would query these
-    /// as hostnames, so they stay on the getaddrinfo path.
-    pub(crate) fn getaddrinfo_only_name(name: &[u8]) -> bool {
-        strings::is_ip_address(name)
-            || strings::contains_char(name, b'%')
-            // inet_aton forms getaddrinfo accepts: `127.1`, `2130706433`,
-            // `0x7f000001`, `0177.0.0.1`.
-            || (name.first().is_some_and(|b| b.is_ascii_digit())
-                && name
-                    .iter()
-                    .all(|b| b.is_ascii_hexdigit() || matches!(*b, b'.' | b'x' | b'X')))
-    }
-
-    /// `hints` bits dns_sd has no equivalent for (AI_V4MAPPED/AI_ALL/…);
-    /// `AI_ADDRCONFIG` maps to `kDNSServiceFlagsSuppressUnusable` and is
-    /// handled here.
-    pub(crate) fn getaddrinfo_only_flags(flags: c_int) -> bool {
-        flags & !netc::AI_ADDRCONFIG != 0
-    }
-
-    /// `kDNSServiceFlagsSuppressUnusable` is the daemon-side `AI_ADDRCONFIG`:
-    /// when both families are requested, a family this host has no routable
-    /// address for is answered locally with `NoSuchRecord` instead of a
-    /// wire query, and `/etc/hosts` names (localhost) are exempt. libinfo's
-    /// getaddrinfo sets it whenever `AI_ADDRCONFIG` is (mdns_module.c).
-    fn addrconfig_flags(protocol: DNSServiceProtocol) -> DNSServiceFlags {
-        if protocol != (PROTOCOL_IPV4 | PROTOCOL_IPV6)
-            || env_var::feature_flag::BUN_FEATURE_FLAG_DISABLE_ADDRCONFIG
-                .get()
-                .unwrap_or(false)
-        {
-            return 0;
-        }
-        FLAGS_SUPPRESS_UNUSABLE
-    }
-
-    /// How long to keep waiting for the other family once one has
-    /// produced answers. libinfo's mdns module bounds the wait for the
-    /// second address family after the first answers rather than sitting out
-    /// the full query timeout, so an A-only name whose AAAA question goes
-    /// unanswered resolves in ~this instead of when the daemon times out.
-    const SECOND_FAMILY_EXTRA_MS: i64 = 2000;
-
-    fn now_ms() -> i64 {
-        let now = bun::timespec::now(bun::TimespecMockMode::AllowMockedTime);
-        now.sec * 1000 + (now.nsec as i64) / 1_000_000
-    }
-
-    /// Per-query state for one `DNSServiceGetAddrInfo` subordinate, shared
-    /// by the JS `dns.lookup` request and the internal connect-path request
-    /// so both apply identical completion rules.
-    #[derive(Default)]
-    pub(crate) struct QueryState {
-        pub(crate) sd_ref: DNSServiceRef,
-        pub(crate) results: bun_dns::ResultList,
-        /// First hard `DNSServiceErrorType` (NoSuchRecord/Timeout are
-        /// per-family negatives, not errors).
-        pub(crate) sd_error: DNSServiceErrorType,
-        /// A family reported `kDNSServiceErr_Timeout`; with no results this
-        /// is a transient EAI_AGAIN, not a definitive EAI_NONAME.
-        saw_timeout: bool,
-        /// The last reply for this request carried `kDNSServiceFlagsMoreComing`
-        /// and no other request's reply has followed it yet: the daemon has
-        /// more of this answer set queued (possibly not yet written). If that
-        /// continuation is never delivered (a stale reply for a deallocated
-        /// subordinate is dropped by the client stub), `kDNSServiceFlagsTimeout`
-        /// still guarantees a final callback that completes the request.
-        awaiting_more: bool,
-        /// `kDNSServiceProtocol_*` bits still awaiting any reply. A family's
-        /// bit is cleared by any callback carrying its `sa_family` (Add,
-        /// NoSuchRecord, or Timeout; the client stub always passes a
-        /// family-tagged sockaddr).
-        pub(crate) pending_proto: DNSServiceProtocol,
-        /// Some family has answered while another is still pending: the
-        /// timestamp starts the `SECOND_FAMILY_EXTRA_MS` early-out.
-        partial_at_ms: Option<i64>,
-        gave_up_on_pending: bool,
-    }
-
-    impl QueryState {
-        pub(crate) fn new(protocol: DNSServiceProtocol) -> Self {
-            Self {
-                pending_proto: protocol,
-                ..Default::default()
-            }
-        }
-
-        /// Absorb one callback from `DNSServiceGetAddrInfo`.
-        /// SAFETY: `address`, if non-null, points at a valid sockaddr of the
-        /// family it declares (guaranteed by dnssd_clientstub for the callback).
-        pub(crate) unsafe fn record_reply(
-            &mut self,
-            flags: DNSServiceFlags,
-            error_code: DNSServiceErrorType,
-            address: *const Sockaddr,
-            ttl: u32,
-        ) {
-            self.awaiting_more = flags & FLAGS_MORE_COMING != 0;
-            // dnssd_clientstub always passes a family-tagged sockaddr for A/AAAA
-            // replies (including NoSuchRecord/Timeout); only PolicyDenied is null.
-            if address.is_null() {
-                if self.sd_error == 0 {
-                    self.sd_error = error_code;
-                }
-                return;
-            }
-            // SAFETY: caller contract.
-            let fam = unsafe { (*address).sa_family } as i32;
-            // Any reply for a family retires its pending bit; whether the
-            // whole answer set has arrived is tracked by `awaiting_more`.
-            self.pending_proto &= !if fam == netc::AF_INET6 {
-                PROTOCOL_IPV6
-            } else {
-                PROTOCOL_IPV4
-            };
-            if error_code == ERR_NO_ERROR && flags & FLAGS_ADD != 0 {
-                self.results.push(GetAddrInfoResult {
-                    // SAFETY: caller contract.
-                    address: unsafe { bun_dns::Address::init_posix(address.cast()) },
-                    ttl: ttl as i32,
-                });
-            } else if error_code == ERR_TIMEOUT {
-                self.saw_timeout = true;
-            } else if error_code != ERR_NO_ERROR
-                && error_code != ERR_NO_SUCH_RECORD
-                && self.sd_error == 0
-            {
-                self.sd_error = error_code;
-            }
-            if !self.results.is_empty() && self.pending_proto != 0 && self.partial_at_ms.is_none() {
-                self.partial_at_ms = Some(now_ms());
-            }
-        }
-
-        pub(crate) fn is_ready(&self) -> bool {
-            self.sd_error != 0
-                || self.gave_up_on_pending
-                || (self.pending_proto == 0 && !self.awaiting_more)
-        }
-
-        /// Deadline for giving up on the still-pending family, if one family
-        /// has already answered.
-        pub(crate) fn early_out_deadline_ms(&self) -> Option<i64> {
-            if self.pending_proto == 0 || self.gave_up_on_pending {
-                return None;
-            }
-            self.partial_at_ms.map(|t| t + SECOND_FAMILY_EXTRA_MS)
-        }
-
-        /// EAI_* status for a completed query with no results.
-        pub(crate) fn empty_status(&self) -> c_int {
-            if self.sd_error != 0 {
-                to_eai(self.sd_error)
-            } else if self.saw_timeout {
-                libc::EAI_AGAIN
-            } else {
-                libc::EAI_NONAME
-            }
-        }
-
-        pub(crate) fn take_results(&mut self) -> bun_dns::ResultList {
-            let mut results = core::mem::take(&mut self.results);
-            // Arrival order across families races (each family is a separate
-            // A/AAAA question upstream); getaddrinfo sorted with the RFC 6724
-            // default policy, which puts IPv6 first when both are configured.
-            results.sort_by_key(|r| r.address.family() != netc::AF_INET6);
-            results
-        }
-    }
-
-    #[derive(Clone, Copy)]
-    pub(crate) enum Inflight {
-        Jsc(*mut GetAddrInfoRequest),
-        Internal(*mut internal::Request),
-    }
-
-    impl Inflight {
-        fn context(&self) -> *mut c_void {
-            match *self {
-                Inflight::Jsc(r) => r.cast(),
-                Inflight::Internal(r) => r.cast(),
-            }
-        }
-
-        /// SAFETY: the request behind `self` is live (pinned in `inflight`).
-        unsafe fn query(&self) -> &mut QueryState {
-            // SAFETY: caller contract; event-loop thread only.
-            unsafe {
-                match *self {
-                    Inflight::Jsc(r) => &mut (*r).backend.as_dns_sd_mut().query,
-                    Inflight::Internal(r) => &mut (*r).libinfo.query,
-                }
-            }
-        }
-    }
-
-    /// One per event loop. Owns the primary `DNSServiceRef` and its `FilePoll`;
-    /// every lookup is a `kDNSServiceFlagsShareConnection` subordinate on it.
-    pub(crate) struct SharedConnection {
-        main_ref: DNSServiceRef,
-        file_poll: NonNull<FilePoll>,
-        ctx: Async::EventLoopCtx,
-        inflight: Vec<Inflight>,
-        /// `context` of the previous reply on the wire. mDNSResponder queues
-        /// all answers from one DNS response contiguously and stamps
-        /// `MoreComing` on a reply whenever another is queued behind it on
-        /// this connection, so a request's contiguous run has ended once a
-        /// reply for a different request follows it.
-        last_ctx: *mut c_void,
-        /// Runtime-heap timer for `SECOND_FAMILY_EXTRA_MS` early-outs (armed
-        /// only on JS threads; the daemon's own timeout is the backstop on the
-        /// HTTP-thread loop).
-        pub(crate) early_out_timer: JsCell<EventLoopTimer>,
-        /// Deadline the timer is currently armed for (0 = disarmed).
-        early_out_armed_for: Cell<i64>,
-    }
-
-    bun_event_loop::impl_timer_owner!(SharedConnection; from_early_out_timer_ptr => early_out_timer);
-
-    thread_local! {
-        static SHARED: Cell<*mut SharedConnection> = const { Cell::new(ptr::null_mut()) };
-    }
-
-    impl SharedConnection {
-        /// Lazily create the shared connection for this event loop. Returns
-        /// `None` if mDNSResponder is unreachable; the caller falls back to
-        /// blocking `getaddrinfo` on the work pool.
-        pub(crate) fn get(ctx: Async::EventLoopCtx) -> Option<*mut Self> {
-            let existing = SHARED.get();
-            if !existing.is_null() {
-                return Some(existing);
-            }
-            let mut main_ref: DNSServiceRef = ptr::null_mut();
-            // SAFETY: FFI; `main_ref` is stack-local.
-            let err = unsafe { DNSServiceCreateConnection(&raw mut main_ref) };
-            if err != ERR_NO_ERROR || main_ref.is_null() {
-                bun_output::scoped_log!(dns, "DNSServiceCreateConnection failed: {}", err);
-                return None;
-            }
-            // SAFETY: FFI; `main_ref` is a live ref just returned by CreateConnection.
-            let fd = sys::Fd::from_native(unsafe { DNSServiceRefSockFD(main_ref) });
-            let this = bun_core::heap::into_raw(Box::new(Self {
-                main_ref,
-                file_poll: NonNull::dangling(),
-                ctx,
-                inflight: Vec::new(),
-                last_ctx: ptr::null_mut(),
-                early_out_timer: JsCell::new(EventLoopTimer::init_paused(
-                    EventLoopTimerTag::DnsSdConnection,
-                )),
-                early_out_armed_for: Cell::new(0),
-            }));
-            let poll_ptr = FilePoll::init(
-                ctx,
-                fd,
-                Default::default(),
-                Async::Owner::new(
-                    Async::posix_event_loop::poll_tag::GET_ADDR_INFO_REQUEST,
-                    this.cast(),
-                ),
-            );
-            // SAFETY: FilePoll::init returns a live pool slot; exclusive here.
-            let poll = unsafe { &mut *poll_ptr };
-            let rc = poll.register_with_fd(
-                // SAFETY: the event loop is live for the caller's lifetime.
-                unsafe { ctx.platform_event_loop() },
-                Async::PollKind::Readable,
-                Async::posix_event_loop::OneShotFlag::None,
-                fd,
-            );
-            if rc.is_err() {
-                // SAFETY: `poll_ptr` is the fresh hive slot; `main_ref` is the
-                // live connection ref; `this` is the just-allocated box.
-                unsafe {
-                    (*poll_ptr).deinit();
-                    DNSServiceRefDeallocate(main_ref);
-                    drop(bun_core::heap::take(this));
-                }
-                return None;
-            }
-            // SAFETY: `this` is the freshly heap-allocated box; exclusive here.
-            unsafe { (*this).file_poll = NonNull::new(poll_ptr).unwrap() };
-            SHARED.set(this);
-            Some(this)
-        }
-
-        /// Start a subordinate `DNSServiceGetAddrInfo` on the shared
-        /// connection and register it as in-flight. Bumps the FilePoll's
-        /// keep-alive on the 0→1 transition so the process stays up while a
-        /// lookup is pending. Returns the sub-ref on success; `None` if the
-        /// daemon rejected the request.
-        pub(crate) fn start(
-            this: *mut Self,
-            owner: Inflight,
-            protocol: DNSServiceProtocol,
-            hostname: *const c_char,
-            callback: GetAddrInfoReply,
-            context: *mut c_void,
-        ) -> Option<DNSServiceRef> {
-            // SAFETY: `this` is the live heap-allocated connection from `get()`.
-            let mut sub: DNSServiceRef = unsafe { (*this).main_ref };
-            // SAFETY: FFI; `hostname` is NUL-terminated and dns_sd copies it
-            // before returning. `sub` holds a copy of the primary ref as
-            // required for ShareConnection.
-            let err = unsafe {
-                DNSServiceGetAddrInfo(
-                    &raw mut sub,
-                    FLAGS_SHARE_CONNECTION
-                        | FLAGS_TIMEOUT
-                        | FLAGS_RETURN_INTERMEDIATES
-                        | addrconfig_flags(protocol),
-                    0,
-                    protocol,
-                    hostname,
-                    callback,
-                    context,
-                )
-            };
-            if err != ERR_NO_ERROR {
-                bun_output::scoped_log!(dns, "DNSServiceGetAddrInfo failed: {}", err);
-                return None;
-            }
-            // SAFETY: `this` is live; event-loop-thread-only so no data race.
-            unsafe {
-                if (*this).inflight.is_empty() {
-                    (*(*this).file_poll.as_ptr()).enable_keeping_process_alive((*this).ctx);
-                }
-                (*this).inflight.push(owner);
-            }
-            Some(sub)
-        }
-
-        /// Called first thing from each request's reply callback (which only
-        /// fires inside `on_readable`'s `DNSServiceProcessResult`). When the
-        /// wire moves on to a different request, the previous request's
-        /// contiguous run of replies has ended, so its `awaiting_more` (from
-        /// a `MoreComing` that referred to *this* reply) is cleared.
-        pub(crate) fn note_reply(context: *mut c_void) {
-            let this = SHARED.get();
-            if this.is_null() {
-                return;
-            }
-            // SAFETY: `this` is the live connection whose `on_readable` is on
-            // the stack; the callback runs on the same (event-loop) thread and
-            // `on_readable` holds no borrow of `*this` across ProcessResult.
-            unsafe {
-                let prev = (*this).last_ctx;
-                if !prev.is_null() && prev != context {
-                    for inf in (*this).inflight.iter() {
-                        if inf.context() == prev {
-                            inf.query().awaiting_more = false;
-                        }
-                    }
-                }
-                (*this).last_ctx = context;
-            }
-        }
-
-        /// kqueue `EVFILT_READ` fired on the shared socket. Drains all pending
-        /// daemon replies (dnssd's `DNSServiceProcessResult` loops until the
-        /// socket would block, firing each sub-ref's callback inline), then
-        /// completes every query whose answer set is complete.
-        pub(crate) fn on_readable(this: *mut Self) {
-            // `finish()` resolves promises; keep one enter/exit scope open so
-            // the microtask drain (and any JS it runs, which may re-enter
-            // `start()` or tear this connection down) happens after this
-            // function has stopped touching `*this`.
-            // SAFETY: `get_or_null` returns the current thread's VM or null
-            // (the internal connect path also runs on the HTTP thread's loop,
-            // where there is no VM and completion never touches JS).
-            let _exit =
-                VirtualMachine::get_or_null().map(|vm| unsafe { (*vm).enter_event_loop_scope() });
-
-            // SAFETY: `this` is the live heap-allocated connection the FilePoll
-            // was registered with; exclusive on the event-loop thread. The
-            // callbacks fired inside `DNSServiceProcessResult` reach the
-            // per-request state via `context`, not via `*this`. Each borrow of
-            // `*this` below ends before any `finish()` runs.
-            let rc = unsafe { DNSServiceProcessResult((*this).main_ref) };
-            let ready: Vec<Inflight> = if rc != ERR_NO_ERROR {
-                bun_output::scoped_log!(dns, "DNSServiceProcessResult: {}", rc);
-                // SAFETY: as above; `take` leaves an empty Vec in place.
-                unsafe { core::mem::take(&mut (*this).inflight) }
-            } else {
-                let mut ready = Vec::new();
-                // SAFETY: as above; each `*mut` in `inflight` is a live heap
-                // request pinned there until finished.
-                let inflight = unsafe { &mut (*this).inflight };
-                let mut i = 0;
-                while i < inflight.len() {
-                    // SAFETY: each entry is a live heap request pinned in `inflight`.
-                    let is_ready = unsafe { inflight[i].query().is_ready() };
-                    if is_ready {
-                        ready.push(inflight.swap_remove(i));
-                    } else {
-                        i += 1;
-                    }
-                }
-                if inflight.is_empty() {
-                    // SAFETY: `file_poll` is the hive slot set in `get()`.
-                    unsafe {
-                        (*(*this).file_poll.as_ptr()).disable_keeping_process_alive((*this).ctx)
-                    };
-                }
-                ready
-            };
-            if rc != ERR_NO_ERROR {
-                // The primary is defunct (daemon died or socket went bad).
-                // Detach it first so a lookup started from a rejection handler
-                // reconnects, fail the subordinates while the primary is still
-                // alive (deallocating the parent frees them implicitly, and a
-                // later deallocate would double-free), then destroy it.
-                let detached = SHARED.replace(ptr::null_mut());
-                debug_assert!(core::ptr::eq(detached, this));
-                for inf in ready {
-                    Self::finish(inf, Some(rc));
-                }
-                // SAFETY: `this` is detached and its `inflight` was drained
-                // above; nothing else references it.
-                unsafe { Self::destroy(this) };
-            } else {
-                Self::arm_early_out(this);
-                for inf in ready {
-                    Self::finish(inf, None);
-                }
-            }
-        }
-
-        /// Arm the runtime timer for the nearest second-family early-out
-        /// deadline among in-flight queries, if any. No-op off the JS thread
-        /// (no timer heap there); the daemon timeout still bounds the wait.
-        fn arm_early_out(this: *mut Self) {
-            if VirtualMachine::get_or_null().is_none() {
-                return;
-            }
-            // SAFETY: `this` is the live connection; event-loop thread only.
-            unsafe {
-                let mut min_deadline: Option<i64> = None;
-                for inf in (*this).inflight.iter() {
-                    if let Some(d) = inf.query().early_out_deadline_ms() {
-                        min_deadline = Some(min_deadline.map_or(d, |m| m.min(d)));
-                    }
-                }
-                let Some(deadline) = min_deadline else {
-                    return;
-                };
-                let armed = (*this).early_out_armed_for.get();
-                if armed != 0 && armed <= deadline {
-                    return;
-                }
-                let now = bun::timespec::now(bun::TimespecMockMode::AllowMockedTime);
-                let delay_ms = (deadline - now_ms()).max(1);
-                let next = now.add_ms(delay_ms);
-                let timer = &(*this).early_out_timer;
-                let state = crate::jsc_hooks::runtime_state();
-                if timer.get().state == EventLoopTimerState::ACTIVE {
-                    (*state).timer.remove(timer.as_ptr());
-                }
-                timer.with_mut(|t| {
-                    t.next = ElTimespec {
-                        sec: next.sec,
-                        nsec: next.nsec,
-                    }
-                });
-                (*state).timer.insert(timer.as_ptr());
-                (*this).early_out_armed_for.set(deadline);
-            }
-        }
-
-        /// Runtime-timer fire (via `dispatch.rs`): complete queries that have
-        /// had one family's answers for `SECOND_FAMILY_EXTRA_MS` while the
-        /// other family stayed silent.
-        ///
-        /// # Safety
-        /// `this` must be the live connection whose `early_out_timer` fired.
-        pub(crate) unsafe fn on_early_out(this: *mut Self) {
-            // See `on_readable`: keep one scope open across `finish()`.
-            let _exit =
-                VirtualMachine::get_or_null().map(|vm| unsafe { (*vm).enter_event_loop_scope() });
-            // SAFETY: caller contract; borrows of `*this` end before any
-            // `finish()`.
-            unsafe {
-                (*this).early_out_armed_for.set(0);
-                let now = now_ms();
-                let mut ready = Vec::new();
-                let inflight = &mut (*this).inflight;
-                let mut i = 0;
-                while i < inflight.len() {
-                    let q = inflight[i].query();
-                    if q.early_out_deadline_ms()
-                        .is_some_and(|d| d <= now && !q.results.is_empty())
-                    {
-                        q.gave_up_on_pending = true;
-                        ready.push(inflight.swap_remove(i));
-                    } else {
-                        i += 1;
-                    }
-                }
-                if inflight.is_empty() {
-                    (*(*this).file_poll.as_ptr()).disable_keeping_process_alive((*this).ctx);
-                }
-                Self::arm_early_out(this);
-                for inf in ready {
-                    Self::finish(inf, None);
-                }
-            }
-        }
-
-        /// Free a detached connection with no in-flight sub-refs.
-        /// SAFETY: `this` must be the live heap connection, already removed
-        /// from `SHARED`, with `inflight` empty.
-        unsafe fn destroy(this: *mut Self) {
-            // SAFETY: caller contract; `deinit` returns the hive slot.
-            unsafe {
-                debug_assert!((*this).inflight.is_empty());
-                if (*this).early_out_timer.get().state == EventLoopTimerState::ACTIVE {
-                    if VirtualMachine::get_or_null().is_some() {
-                        (*crate::jsc_hooks::runtime_state())
-                            .timer
-                            .remove((*this).early_out_timer.as_ptr());
-                    }
-                }
-                (*(*this).file_poll.as_ptr()).deinit();
-                DNSServiceRefDeallocate((*this).main_ref);
-                drop(bun_core::heap::take(this));
-            }
-        }
-
-        /// `force_err` (defunct connection / VM teardown) discards any
-        /// partial results so completion always takes the error path —
-        /// resolving inside teardown would drain microtasks and can
-        /// re-enter DNS after the resolver is gone.
-        fn finish(inf: Inflight, force_err: Option<DNSServiceErrorType>) {
-            // SAFETY: each `*mut` is a live heap request just removed from
-            // `inflight`; its `sd_ref` is the live subordinate to be released.
-            unsafe {
-                match inf {
-                    Inflight::Jsc(r) => {
-                        let q = &mut (*r).backend.as_dns_sd_mut().query;
-                        DNSServiceRefDeallocate(q.sd_ref);
-                        if let Some(e) = force_err {
-                            q.results.clear();
-                            q.sd_error = e;
-                        }
-                        GetAddrInfoRequest::complete_dns_sd(r);
-                    }
-                    Inflight::Internal(r) => {
-                        let q = &mut (*r).libinfo.query;
-                        DNSServiceRefDeallocate(q.sd_ref);
-                        if let Some(e) = force_err {
-                            q.results.clear();
-                            q.sd_error = e;
-                        }
-                        internal::dns_sd_complete(r);
-                    }
-                }
-            }
-        }
-
-        /// Tear down the shared connection if one exists on this thread.
-        /// Called from VM teardown so the fd and FilePoll are released while
-        /// the event loop is still live. In-flight requests are failed
-        /// (mirroring c-ares' `EDESTRUCTION` on channel destroy) so their
-        /// heap allocations are reclaimed before the VM goes.
-        pub(crate) fn close_for_terminate() {
-            let this = SHARED.replace(ptr::null_mut());
-            if this.is_null() {
-                return;
-            }
-            // SAFETY: `this` is the live heap-allocated connection just
-            // detached from `SHARED`; exclusive on this (event-loop) thread.
-            // Subordinates are failed (deallocating them) before the parent.
-            unsafe {
-                while let Some(inf) = (*this).inflight.pop() {
-                    Self::finish(inf, Some(ERR_DEFUNCT_CONNECTION));
-                }
-                Self::destroy(this);
-            }
-        }
-    }
-
-    pub(crate) fn lookup(
-        this: &Resolver,
-        query: &GetAddrInfo,
-        global_this: &JSGlobalObject,
-    ) -> JSValue {
-        bun_core::Environment::only_mac();
-
-        if getaddrinfo_only_flags(query.options.flags) || getaddrinfo_only_name(query.name.as_ref())
-        {
-            return lib_c::lookup(this, query, global_this);
-        }
-
-        let key = get_addr_info_request::PendingCacheKey::init(query);
-        let cache =
-            this.get_or_put_into_pending_cache(&key, PendingCacheField::PendingHostCacheNative);
-
-        if let CacheHit::Inflight(inflight) = cache {
-            let dns_lookup = DNSLookup::init(this.as_ctx_ptr(), global_this);
-            // SAFETY: inflight points into resolver's HiveArray buffer
-            unsafe { (*inflight).append(dns_lookup) };
-            // SAFETY: `dns_lookup` was just heap-allocated by `DNSLookup::init`.
-            return unsafe { (*dns_lookup).promise.value() };
-        }
-
-        let Some(shared) = SharedConnection::get(js_event_loop_ctx()) else {
-            if let CacheHit::New(new) = cache {
-                this.pending_host_cache_native.with_mut(|c| {
-                    // SAFETY: `new` is the freshly-allocated HiveArray slot
-                    // returned by `get_or_put_into_pending_cache`; no other
-                    // token for it is outstanding.
-                    unsafe { c.put(new) };
-                });
-            }
-            return lib_c::lookup(this, query, global_this);
-        };
-
-        let protocol = protocol_for_family(query.options.family);
-        let request = GetAddrInfoRequest::init(
-            cache,
-            get_addr_info_request::Backend::DnsSd(get_addr_info_request::BackendDnsSd::new(
-                protocol,
-            )),
-            Some(this.as_ctx_ptr()),
-            query,
-            global_this,
-            PendingCacheField::PendingHostCacheNative,
-        );
-        // SAFETY: request was just heap-allocated in init() and is exclusively owned here.
-        let promise_value = unsafe { (*request).head.promise.value() };
-
-        let name_z = bun::ZBox::from_bytes(query.name.as_ref());
-        let Some(sub) = SharedConnection::start(
-            shared,
-            Inflight::Jsc(request),
-            protocol,
-            name_z.as_ptr().cast::<c_char>(),
-            GetAddrInfoRequest::dns_sd_reply,
-            request.cast::<c_void>(),
-        ) else {
-            // SAFETY: request is exclusively owned; dns_sd never accepted it.
-            unsafe {
-                if (*request).cache.pending_cache() {
-                    let pos = (*request).cache.pos_in_pending();
-                    this.pending_host_cache_native.with_mut(|c| {
-                        let slot = c.ptr_at(pos as usize);
-                        // SAFETY: `pos` was alloc'd; no other token outstanding.
-                        c.put(slot);
-                    });
-                }
-                DNSLookup::destroy(&raw mut (*request).head);
-                drop(bun_core::heap::take(request));
-            }
-            return lib_c::lookup(this, query, global_this);
-        };
-
-        // SAFETY: request is live (heap-allocated) and exclusively accessed here.
-        unsafe { (*request).backend.as_dns_sd_mut().query.sd_ref = sub };
-        this.request_sent(this.vm());
-
-        promise_value
-    }
-}
+#[path = "dns_sd.rs"]
+pub(crate) mod dns_sd;
 
 // ──────────────────────────────────────────────────────────────────────────
 // LibC (blocking getaddrinfo on a worker thread; non-Windows)
@@ -1953,10 +1189,8 @@ impl GetAddrInfoRequest {
         request
     }
 
-    /// Per-address callback from `DNSServiceGetAddrInfo`. Runs synchronously
-    /// inside `DNSServiceProcessResult` on the JS thread; only accumulates
-    /// state. Completion happens in `SharedConnection::on_readable` after the
-    /// process-result call returns.
+    /// Reply callback (fires inside `DNSServiceProcessResult`); only records
+    /// state — completion happens after the drain in `on_readable`.
     ///
     /// # Safety
     /// `context` must be the `*mut GetAddrInfoRequest` registered via
@@ -1987,8 +1221,7 @@ impl GetAddrInfoRequest {
         }
     }
 
-    /// Complete a dns_sd-backed request after `SharedConnection::on_readable`
-    /// observed readiness.
+    /// Complete a dns_sd-backed request once its query is ready.
     ///
     /// # Safety
     /// `this` must be the live heap `GetAddrInfoRequest`; consumed on every path.
@@ -3378,24 +2611,23 @@ pub mod internal {
         };
 
         let protocol = dns_sd::protocol_for_hints(&get_hints());
-        let Some(sub) = dns_sd::SharedConnection::start(
-            shared,
-            dns_sd::Inflight::Internal(req),
-            protocol,
-            host.as_ptr().cast::<c_char>(),
-            dns_sd_reply,
-            req.cast::<c_void>(),
-        ) else {
-            return false;
-        };
-
         // SAFETY: `req` is the live heap-allocated request owned by the caller.
         unsafe {
             (*req).libinfo = MacAsyncDNS {
                 query: dns_sd::QueryState::new(protocol),
             };
-            (*req).libinfo.query.sd_ref = sub;
         }
+        let Some(sub) = shared.start(
+            dns_sd::Inflight::Internal(req),
+            protocol,
+            host,
+            dns_sd_reply,
+            req.cast::<c_void>(),
+        ) else {
+            return false;
+        };
+        // SAFETY: as above; the subordinate is only stored here.
+        unsafe { (*req).libinfo.query.sd_ref = sub };
 
         true
     }
@@ -3727,12 +2959,12 @@ pub mod internal {
                     jsc::EventLoopHandle::from_tag_ptr(tag, ptr)
                 };
                 let res = lookup_dns_sd(req, handle);
-                bun_output::scoped_log!(
-                    dns,
-                    "getaddrinfo({}) = cache miss (dns_sd)",
-                    bstr::BStr::new(host.map(|h| h.as_bytes()).unwrap_or(b""))
-                );
                 if res {
+                    bun_output::scoped_log!(
+                        dns,
+                        "getaddrinfo({}) = cache miss (dns_sd)",
+                        bstr::BStr::new(host.map(|h| h.as_bytes()).unwrap_or(b""))
+                    );
                     return Some(req);
                 }
                 // if dns_sd was unavailable, fall back to the work pool
